+++ src/input_context/input_condition/block_by.rs
@@ -91,10 +91,10 @@ mod tests {
     fn block() {
         let mut condition = BlockBy::<DummyAction>::default();
         let mut action = ActionData::new::<DummyAction>();
-        action.set_state(ActionState::Fired);
+        let time = Time::default();
+        action.update(&time, ActionState::Fired, true);
         let mut actions = ActionsData::default();
         actions.insert_action::<DummyAction>(action);
-        let time = Time::default();
 
         assert_eq!(
             condition.evaluate(&actions, &time, true.into()),